+++ modules/AWSPowerShell/Cmdlets/EC2/Basic/New-EC2SnapshotBatch-Cmdlet.cs
@@ -105,7 +105,14 @@ namespace Amazon.PowerShell.Cmdlets.EC2
         /// <para>The instance to specify which volumes should be snapshotted.</para>
         /// </para>
         /// </summary>
+        #if !MODULAR
         [System.Management.Automation.Parameter(Position = 0, ValueFromPipelineByPropertyName = true, ValueFromPipeline = true)]
+        #else
+        [System.Management.Automation.Parameter(Position = 0, ValueFromPipelineByPropertyName = true, ValueFromPipeline = true, Mandatory = true)]
+        [System.Management.Automation.AllowEmptyString]
+        [System.Management.Automation.AllowNull]
+        #endif
+        [Amazon.PowerShell.Common.AWSRequiredParameter]
         public System.String InstanceSpecification_InstanceId { get; set; }
         #endregion
         
@@ -206,6 +213,12 @@ namespace Amazon.PowerShell.Cmdlets.EC2
                 context.InstanceSpecification_ExcludeDataVolumeId = new List<System.String>(this.InstanceSpecification_ExcludeDataVolumeId);
             }
             context.InstanceSpecification_InstanceId = this.InstanceSpecification_InstanceId;
+            #if MODULAR
+            if (this.InstanceSpecification_InstanceId == null && ParameterWasBound(nameof(this.InstanceSpecification_InstanceId)))
+            {
+                WriteWarning("You are passing $null as a value for parameter InstanceSpecification_InstanceId which is marked as required. In case you believe this parameter was incorrectly marked as required, report this by opening an issue at https://github.com/aws/aws-tools-for-powershell/issues.");
+            }
+            #endif
             context.OutpostArn = this.OutpostArn;
             if (this.TagSpecification != null)
             {