+++ modules/AWSPowerShell/Cmdlets/EC2/Basic/Remove-EC2LocalGatewayRoute-Cmdlet.cs
@@ -46,15 +46,19 @@ namespace Amazon.PowerShell.Cmdlets.EC2
         /// <para>The CIDR range for the route. This must match the CIDR for the route exactly.</para>
         /// </para>
         /// </summary>
-        #if !MODULAR
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
-        #else
-        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true, Mandatory = true)]
-        [System.Management.Automation.AllowEmptyString]
-        [System.Management.Automation.AllowNull]
-        #endif
-        [Amazon.PowerShell.Common.AWSRequiredParameter]
         public System.String DestinationCidrBlock { get; set; }
+        #endregion
+        
+        #region Parameter DestinationPrefixListId
+        /// <summary>
+        /// <para>
+        /// <para> Use a prefix list in place of <code>DestinationCidrBlock</code>. You cannot use <code>DestinationPrefixListId</code>
+        /// and <code>DestinationCidrBlock</code> in the same request. </para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public System.String DestinationPrefixListId { get; set; }
         #endregion
         
         #region Parameter LocalGatewayRouteTableId
@@ -137,12 +141,7 @@ namespace Amazon.PowerShell.Cmdlets.EC2
             }
             #pragma warning restore CS0618, CS0612 //A class member was marked with the Obsolete attribute
             context.DestinationCidrBlock = this.DestinationCidrBlock;
-            #if MODULAR
-            if (this.DestinationCidrBlock == null && ParameterWasBound(nameof(this.DestinationCidrBlock)))
-            {
-                WriteWarning("You are passing $null as a value for parameter DestinationCidrBlock which is marked as required. In case you believe this parameter was incorrectly marked as required, report this by opening an issue at https://github.com/aws/aws-tools-for-powershell/issues.");
-            }
-            #endif
+            context.DestinationPrefixListId = this.DestinationPrefixListId;
             context.LocalGatewayRouteTableId = this.LocalGatewayRouteTableId;
             #if MODULAR
             if (this.LocalGatewayRouteTableId == null && ParameterWasBound(nameof(this.LocalGatewayRouteTableId)))
@@ -169,6 +168,10 @@ namespace Amazon.PowerShell.Cmdlets.EC2
             if (cmdletContext.DestinationCidrBlock != null)
             {
                 request.DestinationCidrBlock = cmdletContext.DestinationCidrBlock;
+            }
+            if (cmdletContext.DestinationPrefixListId != null)
+            {
+                request.DestinationPrefixListId = cmdletContext.DestinationPrefixListId;
             }
             if (cmdletContext.LocalGatewayRouteTableId != null)
             {
@@ -236,6 +239,7 @@ namespace Amazon.PowerShell.Cmdlets.EC2
         internal partial class CmdletContext : ExecutorContext
         {
             public System.String DestinationCidrBlock { get; set; }
+            public System.String DestinationPrefixListId { get; set; }
             public System.String LocalGatewayRouteTableId { get; set; }
             public System.Func<Amazon.EC2.Model.DeleteLocalGatewayRouteResponse, RemoveEC2LocalGatewayRouteCmdlet, object> Select { get; set; } =
                 (response, cmdlet) => response.Route;