+++ modules/AWSPowerShell/Cmdlets/EC2/Basic/Search-EC2LocalGatewayRoute-Cmdlet.cs
@@ -43,7 +43,7 @@ namespace Amazon.PowerShell.Cmdlets.EC2
         #region Parameter Filter
         /// <summary>
         /// <para>
-        /// <para>One or more filters.</para><ul><li><para><code>route-search.exact-match</code> - The exact match of the specified filter.</para></li><li><para><code>route-search.longest-prefix-match</code> - The longest prefix that matches
+        /// <para>One or more filters.</para><ul><li><para><code>prefix-list-id</code> - The ID of the prefix list.</para></li><li><para><code>route-search.exact-match</code> - The exact match of the specified filter.</para></li><li><para><code>route-search.longest-prefix-match</code> - The longest prefix that matches
         /// the route.</para></li><li><para><code>route-search.subnet-of-match</code> - The routes with a subnet that match the
         /// specified CIDR filter.</para></li><li><para><code>route-search.supernet-of-match</code> - The routes with a CIDR that encompass
         /// the CIDR filter. For example, if you have 10.0.1.0/29 and 10.0.1.0/31 routes in your